backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
python-multipart>=0.0.6

pinecone-client>=2.2.0

basic-pitch>=0.4.0
tensorflow>=2.11.0,<2.15.0

librosa>=0.10.1
soundfile>=0.12.1
numpy>=1.24.0,<1.25.0  # Pinned for TensorFlow compatibility
scipy>=1.11.0
resampy>=0.4.2

scikit-learn>=1.3.0
mir_eval>=0.7

python-dotenv>=1.0.0
requests>=2.31.0
pydantic>=2.0.0
pandas>=2.0.0
pathlib
typing-extensions

pytest>=7.0.0
pytest-asyncio>=0.21.0